% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homebreak.R
\name{home_break}
\alias{home_break}
\title{Describes your home surf break and favorite surfer}
\usage{
home_break(fave_break, visits, fave_surfer)
}
\arguments{
\item{fave_break}{a character containing the name of the users home break}

\item{visits}{a number containing how often user visits home break}

\item{fave_surfer}{a character containing a name of the users favorite surfer}
}
\value{

}
\description{
Contains a function, homebreak, that prints a statement about the users home break, how often they surf and fave surfer
}
